backend/requirements.txt
alembic==1.13.3
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==3.0.1
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.3.4
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
qrcode==8.0
requests==2.32.3
simple-websocket==1.1.0
SQLAlchemy==2.0.36
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
wsproto==1.2.0
